requirements.txt
streamlit
groq
pymupdf
pywebview
pyinstaller